+++ src/gpio/hal_1.rs
@@ -9,11 +9,11 @@ use embedded_hal::digital::{
 };
 
 // Implementations for `Pin`
-impl<const P: usize, const N: u8, MODE> ErrorType for Pin<P, N, MODE> {
+impl<const P: usize, const N: usize, MODE> ErrorType for Pin<P, N, MODE> {
     type Error = Infallible;
 }
 
-impl<const P: usize, const N: u8, MODE> OutputPin for Pin<P, N, Output<MODE>> {
+impl<const P: usize, const N: usize, MODE> OutputPin for Pin<P, N, Output<MODE>> {
     #[inline(always)]
     fn set_high(&mut self) -> Result<(), Self::Error> {
         self.set_high();
@@ -29,7 +29,7 @@ impl<const P: usize, const N: u8, MODE> OutputPin for Pin<P, N, Output<MODE>> {
 
 // TODO (alepez) StatefulOutputPin was implemented in stm32. It let the user know the state of output pin
 //   tc37x output register are read only, so it's only possible by remembering the state.
-impl<const P: usize, const N: u8, MODE> StatefulOutputPin for Pin<P, N, Output<MODE>> {
+impl<const P: usize, const N: usize, MODE> StatefulOutputPin for Pin<P, N, Output<MODE>> {
     #[inline(always)]
     fn is_set_high(&self) -> Result<bool, Self::Error> {
         todo!()
@@ -41,7 +41,7 @@ impl<const P: usize, const N: u8, MODE> StatefulOutputPin for Pin<P, N, Output<M
     }
 }
 
-impl<const P: usize, const N: u8, MODE> ToggleableOutputPin for Pin<P, N, Output<MODE>> {
+impl<const P: usize, const N: usize, MODE> ToggleableOutputPin for Pin<P, N, Output<MODE>> {
     #[inline(always)]
     fn toggle(&mut self) -> Result<(), Self::Error> {
         self.toggle();
@@ -49,7 +49,7 @@ impl<const P: usize, const N: u8, MODE> ToggleableOutputPin for Pin<P, N, Output
     }
 }
 
-impl<const P: usize, const N: u8, MODE> InputPin for Pin<P, N, MODE>
+impl<const P: usize, const N: usize, MODE> InputPin for Pin<P, N, MODE>
 where
     MODE: marker::Readable,
 {
@@ -173,11 +173,11 @@ where
 }
 
 // Implementations for `DynamicPin
-impl<const P: usize, const N: u8> ErrorType for DynamicPin<P, N> {
+impl<const P: usize, const N: usize> ErrorType for DynamicPin<P, N> {
     type Error = PinModeError;
 }
 
-impl<const P: usize, const N: u8> OutputPin for DynamicPin<P, N> {
+impl<const P: usize, const N: usize> OutputPin for DynamicPin<P, N> {
     fn set_high(&mut self) -> Result<(), Self::Error> {
         self.set_high()
     }
@@ -186,7 +186,7 @@ impl<const P: usize, const N: u8> OutputPin for DynamicPin<P, N> {
     }
 }
 
-impl<const P: usize, const N: u8> InputPin for DynamicPin<P, N> {
+impl<const P: usize, const N: usize> InputPin for DynamicPin<P, N> {
     fn is_high(&self) -> Result<bool, Self::Error> {
         self.is_high()
     }
